here is the spawmanager.cs file
It runs on my version by only spawning the one available child at random location. My file takes in a controller input and makes a random ordering of the mom. it then spawns the children at random available locations.  I m currently trying to update my version to make it as updated as possible but we need the multiplayer to fully work first.